update comments and move mock types into a mock module (only accessible in the super module)
 - mocks can be used in both types.rs and verifier.rs without decreasing codecov
